requirements.txt
# Core dependencies
ccxt
python-dotenv==1.0.0
rich
pandas
numpy
scipy
pyarrow

# ML/Training dependencies
scikit-learn
xgboost
lightgbm
optuna
joblib
# hmmlearn

# Dev/Test
pytest
